requirements.txt
streamlit
python-docx
openai
pillow<10.0.0